Fix plugin converting user links to videos
An issue in the regex intentionally placed for the simultaneous parsing of youtu
be.com/watch?v= and youtu.be/ links caused links in the format youtube.com/user/
somebody and youtube.com/somebody to wrongly show up as videos. This made going
to the links impossible. This commit separates the shortUrl into its own regex t
o be matched separately, so the regular URL does not have false positives.